+++ src/frykit/plot/utils.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
+from collections.abc import Callable
 from functools import lru_cache
-from typing import Callable
+from typing import Sequence, overload
 
 import numpy as np
 import shapely
@@ -11,7 +12,6 @@ from numpy.typing import NDArray
 from pyproj import CRS, Transformer
 from shapely.geometry.base import BaseGeometry, BaseMultipartGeometry
 
-from frykit.calc import is_finite
 from frykit.shp.typing import GeometryT, PolygonType
 from frykit.utils import format_type_error
 
@@ -35,16 +35,25 @@ def box_path(x0: float, x1: float, y0: float, y1: float, ccw: bool = True) -> Pa
     return Path(vertices, closed=True)
 
 
-EMPTY_PATH = Path(np.empty((0, 2)))
+EMPTY_PATH = Path(np.empty([0, 2]))
 EMPTY_POLYGON = shapely.Polygon()
 
 
-def _line_string_codes(n: int) -> list[np.uint8]:
-    return [Path.MOVETO, *([Path.LINETO] * (n - 1))]
+def _line_string_codes(n: int) -> NDArray[np.uint8]:
+    codes = np.zeros(n, dtype=np.uint8)
+    codes[0] = Path.MOVETO
+    codes[1:] = Path.LINETO
+
+    return codes
 
 
-def _linear_ring_codes(n: int) -> list[np.uint8]:
-    return [Path.MOVETO, *([Path.LINETO] * (n - 2)), Path.CLOSEPOLY]
+def _linear_ring_codes(n: int) -> NDArray[np.uint8]:
+    codes = np.zeros(n, dtype=np.uint8)
+    codes[0] = Path.MOVETO
+    codes[1:-1] = Path.LINETO
+    codes[-1] = Path.CLOSEPOLY
+
+    return codes
 
 
 def geometry_to_path(geometry: BaseGeometry) -> Path:
@@ -57,7 +66,8 @@ def geometry_to_path(geometry: BaseGeometry) -> Path:
 
     See Also
     --------
-    cartopy.mpl.patch.geos_to_path
+    - cartopy.mpl.patch.geos_to_path
+    - cartopy.mpl.path.shapely_to_path
     """
     if isinstance(geometry, BaseGeometry) and geometry.is_empty:
         return EMPTY_PATH
@@ -76,16 +86,17 @@ def geometry_to_path(geometry: BaseGeometry) -> Path:
 
         # 直接使用 orient 效率太低
         case shapely.Polygon():
-            codes: list[np.uint8] = []
             coords_list: list[NDArray[np.float64]] = []
+            codes_list: list[NDArray[np.uint8]] = []
             for i, linear_ring in enumerate([geometry.exterior, *geometry.interiors]):
                 coords = shapely.get_coordinates(linear_ring)
                 if (i == 0) == linear_ring.is_ccw:
                     coords = coords[::-1]
-                codes.extend(_linear_ring_codes(len(coords)))
                 coords_list.append(coords)
+                codes_list.append(_linear_ring_codes(len(coords)))
 
             vertices = np.vstack(coords_list)
+            codes = np.concatenate(codes_list)
             return Path(vertices, codes)
 
         case BaseMultipartGeometry():
@@ -100,31 +111,24 @@ def path_to_polygon(path: Path) -> PolygonType:
     将 Matplotlib 的 Path 对象转为多边形对象
 
     - 空 Path 对应空多边形
-    - 若坐标含 nan 或 inf，那么整个多边形会变成空多边形。
-    - 要求输入是 geometry_to_path(polygon) 的结果，其它输入不保证结果正确。
+    - 要求输入是 geometry_to_path(polygon) 的结果，其它输入可能产生错误。
 
     See Also
     --------
-    cartopy.mpl.patch.path_to_geos
+    - cartopy.mpl.patch.path_to_geos
+    - cartopy.mpl.path.path_to_shapely
     """
     if len(path.vertices) == 0:  # type: ignore
         return EMPTY_POLYGON
 
-    invalid_flag = False
     collection: list[tuple[shapely.LinearRing, list[shapely.LinearRing]]] = []
     indices = np.nonzero(path.codes == Path.MOVETO)[0][1:]  # type: ignore
     for coords in np.vsplit(path.vertices, indices):
-        if not is_finite(coords):
-            invalid_flag = True
-            continue
         linear_ring = shapely.LinearRing(coords)
         if linear_ring.is_ccw:
-            if not invalid_flag:
-                assert len(collection) > 0
-                collection[-1][1].append(linear_ring)
+            collection[-1][1].append(linear_ring)
         else:
             collection.append((linear_ring, []))
-            invalid_flag = False
 
     polygons = [shapely.Polygon(shell, holes) for shell, holes in collection]
     match len(polygons):
@@ -136,65 +140,57 @@ def path_to_polygon(path: Path) -> PolygonType:
             return shapely.MultiPolygon(polygons)
 
 
-# TODO: shapely.set_coordinates
-def _transform_geometry(
-    geometry: GeometryT, transform: Callable[[NDArray[np.float64]], NDArray[np.float64]]
-) -> GeometryT:
-    """shapely.ops.transform 的修改版，会将变换后坐标含 nan 或 inf 的几何对象设为空对象。"""
-    if isinstance(geometry, BaseGeometry) and geometry.is_empty:
-        return geometry
-
-    match geometry:
-        # Point 可以接受形如 (1, 2) 的坐标
-        case shapely.Point() | shapely.LineString() | shapely.LinearRing():
-            coords = transform(shapely.get_coordinates(geometry))
-            if is_finite(coords):
-                return type(geometry)(coords)
-            else:
-                return type(geometry)()
-
-        case shapely.Polygon():
-            shell = transform(shapely.get_coordinates(geometry.exterior))
-            if not is_finite(shell):
-                return type(geometry)()
-
-            holes: list[NDArray[np.float64]] = []
-            for interior in geometry.interiors:
-                hole = transform(shapely.get_coordinates(interior))
-                if is_finite(hole):
-                    holes.append(hole)
-
-            return type(geometry)(shell, holes)
-
-        case BaseMultipartGeometry():
-            parts = [
-                _transform_geometry(part, transform)
-                for part in geometry.geoms
-                if not part.is_empty
-            ]
-            return type(geometry)(parts)  # type: ignore
-
-        case _:
-            raise TypeError(format_type_error("geometry", geometry, BaseGeometry))
-
-
 @lru_cache
 def make_transformer(crs_from: CRS, crs_to: CRS) -> Transformer:
     """创建 pyproj 的 Transformer 对象"""
     return Transformer.from_crs(crs_from, crs_to, always_xy=True)
 
 
-def project_geometry(geometry: GeometryT, crs_from: CRS, crs_to: CRS) -> GeometryT:
-    """对几何对象做投影"""
-    if crs_from == crs_to:
-        return geometry  # 直接返回不可变对象
-
-    transformer = make_transformer(crs_from, crs_to)
-
-    def transform_coords(coords: NDArray[np.float64]) -> NDArray[np.float64]:
+def _make_transformation(
+    crs_from: CRS, crs_to: CRS
+) -> Callable[[NDArray[np.float64]], NDArray[np.float64]]:
+    # shapely>=2.1.0 且 interleaved=False 时就不需要这个辅助函数了
+    def transformation(coords: NDArray[np.float64]) -> NDArray[np.float64]:
+        transformer = make_transformer(crs_from, crs_to)
         return np.column_stack(transformer.transform(coords[:, 0], coords[:, 1]))
 
-    return _transform_geometry(geometry, transform_coords)
+    return transformation
+
+
+@overload
+def project_geometry(geometry: GeometryT, crs_from: CRS, crs_to: CRS) -> GeometryT: ...
+
+
+@overload
+def project_geometry(
+    geometry: Sequence[GeometryT], crs_from: CRS, crs_to: CRS
+) -> NDArray[np.object_]: ...
+
+
+@overload
+def project_geometry(
+    geometry: NDArray[np.object_], crs_from: CRS, crs_to: CRS
+) -> NDArray[np.object_]: ...
+
+
+def project_geometry(
+    geometry: GeometryT | Sequence[GeometryT] | NDArray[np.object_],
+    crs_from: CRS,
+    crs_to: CRS,
+) -> GeometryT | NDArray[np.object_]:
+    """对一个或一组几何对象做投影。通过直接对坐标数组应用 pyproj 实现。"""
+    geometries = np.array(geometry, dtype=np.object_)
+
+    # 尽管 pyproj.Transformer 也能处理 crs 相等的情况，但还是有明显的性能损失
+    if crs_from != crs_to:
+        transformation = _make_transformation(crs_from, crs_to)
+        geometries = shapely.transform(geometries, transformation)
+
+    # 要模仿 shapely 对数组的处理吗？
+    if geometries.ndim == 0 and not isinstance(geometry, np.ndarray):
+        return geometries.item()
+    else:
+        return geometries
 
 
 def get_axes_extents(ax: _AxesBase) -> tuple[float, float, float, float]: